python/requirements.txt
# python/requirements.txt

confluent-kafka
redis>=4.5.0
snowflake-connector-python
python-dotenv
